Refactored gopro script

+++ render-video-from-gopro-photos.sh
@@ -1,13 +1,24 @@
 #! /usr/bin/env bash
 
 counter=0
-for img in G*.JPG
+for img in G[0-9]*.JPG
 do
-  serial=$(printf "%04d" $counter)
+  serial=$(printf "%06d" $counter)
   new_name="G${serial}.JPG"
+  counter=$(expr $counter + 1)
+  if [ -f "$new_name" ]; then
+    echo "$new_name already exists"
+    continue
+  fi
   echo "$img => $new_name"
   mv $img $new_name
-  counter=$((counter++))
 done
 
-ffmpeg -hwaccel cuda -hwaccel_output_format cuda -r 1 -i G%04d.JPG -c:v h264_nvenc -b:v 5M -pix_fmt cuda output.mp4
+case $(uname -s) in
+  Linux)
+    ffmpeg -hwaccel cuda -hwaccel_output_format cuda -r 1 -i G%06d.JPG -c:v h264_nvenc -b:v 5M -pix_fmt cuda output.mp4
+    ;;
+  Darwin)
+    ffmpeg -r 1 -i G%06d.JPG -c:v h264_videotoolbox -b:v 5M -pix_fmt yuv420p output.mp4
+esac
+
